pyproject.toml
[tool.poetry]
name = "repl_nix_car-fleet"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
SQLAlchemy = "^2.0.19"
django-redirect = "^0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
